PDF/requirements.txt
PyMuPDF
pdfplumber
Pillow
transformers
torch

